staging: Add xdg-session-management protocol
For a variety of cases it's desirable to have a method for negotiating
the restoration of previously-used states for a client's windows. This
helps for e.g., a compositor/client crashing (definitely not due to
bugs) or a backgrounded client deciding to temporarily destroy its
surfaces in order to conserve resources.

This protocol adds a method for managing such negotiation and is loosely
based on the Enlightenment "session recovery" protocol which has been
implemented and functional for roughly two years.

v2: Address review comments
v3: Add xdg_toplevel_session_v1.restore event. Move delete toplevel
    request to xdg_toplevel_session_v1. Remove surface argument from
    xdg_toplevel_session_v1.restored. Remove unused 'invalid_restore'
    (same intent as 'already_mapped'). Add 'invalid_session_id' error.
    Specify size limits and other documentation improvements.

Signed-off-by: Mike Blumenkrantz <michael.blumenkrantz@gmail.com>
Signed-off-by: Jonas Ådahl <jadahl@gmail.com>
Signed-off-by: Carlos Garnacho <carlosg@gnome.org>